requirements.txt
Flask==2.3.2
google-generativeai==0.5.4
Pillow==10.1.0
flask-cors==4.0.0
gunicorn==20.1.0